Extract name registration from ClassServer
