+++ opencog/atoms/base/ClassServer.cc
@@ -41,137 +41,9 @@
 
 using namespace opencog;
 
-ClassServer::ClassServer(void)
-{
-	nTypes = 0;
-	_maxDepth = 0;
-}
-
-static int tmod = 0;
-
-/**
- * Return true if this module has already been loaded.
- *
- * This is a strange kind of hack to allow the cogserver unit
- * tests to be run from the build directory, while *also* having
- * libraries that those unit tests require be installed in the
- * system directories.
- *
- * Essentially, this is a manual implementation of the so-called
- * ODR (One Definition Rule) that the linker-loader uses to avoid
- * loading the same library twice, and extending it so that it works
- * when the "same library" is available in the build directory and
- * in some system directory. These are not actually the "same library"
- * from the point of view of the linker-loader; but they need to be,
- * to get the unit tests to pass.
- *
- * The primary issue is driven by mixed python+scheme unit tests.
- * They will typically load libraries in the build dir, and then,
- * because they use modules, try to load the same lirbries again,
- * this time from the system dir.  This results in the same atom
- * types getting defined twice, and then things go downhill.
- */
-bool ClassServer::beginTypeDecls(const char * mod_name)
-{
-   std::lock_guard<std::mutex> l(type_mutex);
-	std::string mname(mod_name);
-	if (_loaded_modules.end() != _loaded_modules.find(mname))
-		return true;
-
-	_loaded_modules.insert(mname);
-	tmod++;
-	return false;
-}
-
-void ClassServer::endTypeDecls(void)
-{
-	// Valid types are odd-numbered.
-	tmod++;
-}
-
-Type ClassServer::declType(const Type parent, const std::string& name)
-{
-    if (1 != tmod%2)
-        throw InvalidParamException(TRACE_INFO,
-            "Improper type declararition; must call beginTypeDecls() first\n");
-
-    // Check if a type with this name already exists. If it does, then
-    // the second and subsequent calls are to be interpreted as defining
-    // multiple inheritance for this type.  A real-life example is the
-    // GroundedSchemeNode, which inherits from several types.
-    Type type = getType(name);
-    if (type != NOTYPE) {
-        std::lock_guard<std::mutex> l(type_mutex);
-
-        // ... unless someone is accidentally declaring the same type
-        // in some different place. In that case, we throw an error.
-        if (_mod[type] != tmod)
-            throw InvalidParamException(TRACE_INFO,
-                "Type \"%s\" has already been declared (%d)\n",
-                name.c_str(), type);
-
-        inheritanceMap[parent][type] = true;
-
-        Type maxd = 1;
-        setParentRecursively(parent, type, maxd);
-        if (_maxDepth < maxd) _maxDepth = maxd;
-        return type;
-    }
-
-    std::unique_lock<std::mutex> l(type_mutex);
-    // Assign type code and increment type counter.
-    type = nTypes++;
-
-    // Resize inheritanceMap container.
-    inheritanceMap.resize(nTypes);
-    recursiveMap.resize(nTypes);
-    _code2NameMap.resize(nTypes);
-    _mod.resize(nTypes);
-    _atomFactory.resize(nTypes);
-    _validator.resize(nTypes);
-
-    for (auto& bv: inheritanceMap) bv.resize(nTypes, false);
-    for (auto& bv: recursiveMap) bv.resize(nTypes, false);
-
-    inheritanceMap[type][type]   = true;
-    inheritanceMap[parent][type] = true;
-    recursiveMap[type][type]     = true;
-    name2CodeMap[name]           = type;
-    _code2NameMap[type]          = &(name2CodeMap.find(name)->first);
-    _mod[type]                   = tmod;
-
-    Type maxd = 1;
-    setParentRecursively(parent, type, maxd);
-    if (_maxDepth < maxd) _maxDepth = maxd;
-
-    // unlock mutex before sending signal which could call
-    l.unlock();
-
-    // Emit add type signal.
-    _addTypeSignal.emit(type);
-
-    return type;
-}
-
-void ClassServer::setParentRecursively(Type parent, Type type, Type& maxd)
-{
-    if (recursiveMap[parent][type]) return;
-
-    bool incr = false;
-    recursiveMap[parent][type] = true;
-    for (Type i = 0; i < nTypes; ++i) {
-        if ((recursiveMap[i][parent]) and (i != parent)) {
-            incr = true;
-            setParentRecursively(i, type, maxd);
-        }
-    }
-    if (incr) maxd++;
-}
-
-TypeSignal& ClassServer::typeAddedSignal()
-{
-    return _addTypeSignal;
-}
+ClassServer::ClassServer(const NameServer & nameServer):
+	_nameServer(nameServer)
+{}
 
 static Handle validating_factory(const Handle& atom_to_check)
 {
@@ -189,21 +61,26 @@ static Handle validating_factory(const Handle& atom_to_check)
 
 void ClassServer::spliceFactory(Type t, AtomFactory* fact)
 {
+	// N.B. it is too late to synchronize calls with NameServer using a shared mutex.
+	// If registrations of class names interleave with registration of factories,
+	// then the code will not work properly anyway. Before a factory is registered,
+	// the complete class hierarchy must be known.
+	
 	// Find all the factories that belong to parents of this type.
 	std::set<AtomFactory*> ok_to_clobber;
 	ok_to_clobber.insert(validating_factory);
 	for (Type parent=0; parent < t; parent++)
 	{
-		if (recursiveMap[parent][t] and _atomFactory[parent])
+		if (_nameServer.isAncestor(parent, t) and _atomFactory[parent])
 			ok_to_clobber.insert(_atomFactory[parent]);
 	}
 
 	// Set the factory for all children of this type.  Be careful
 	// not to clobber any factories that might have been previously
 	// declared.
-	for (Type chi=t; chi < nTypes; chi++)
+	for (Type chi=t; chi < _nameServer.getNumberOfClasses(); chi++)
 	{
-		if (recursiveMap[t][chi] and
+		if (_nameServer.isAncestor(t, chi) and
 		    (nullptr == _atomFactory[chi] or
 		    ok_to_clobber.end() != ok_to_clobber.find(_atomFactory[chi])))
 		{
@@ -214,31 +91,33 @@ void ClassServer::spliceFactory(Type t, AtomFactory* fact)
 
 void ClassServer::addFactory(Type t, AtomFactory* fact)
 {
-	std::unique_lock<std::mutex> l(type_mutex);
+	std::unique_lock<std::mutex> l(factory_mutex);
+	_atomFactory.resize(_nameServer.getNumberOfClasses());
 	spliceFactory(t, fact);
 }
 
 void ClassServer::addValidator(Type t, Validator* checker)
 {
-	std::unique_lock<std::mutex> l(type_mutex);
+	std::unique_lock<std::mutex> l(factory_mutex);
+	_validator.resize(_nameServer.getNumberOfClasses());
 	_validator[t] = checker;
 	spliceFactory(t, validating_factory);
 
-	for (Type chi=t; chi < nTypes; chi++)
+	for (Type chi=t; chi < _nameServer.getNumberOfClasses(); chi++)
 	{
-		if (recursiveMap[t][chi])
+		if (_nameServer.isAncestor(t, chi))
 			_validator[chi] = checker;
 	}
 }
 
 ClassServer::AtomFactory* ClassServer::getFactory(Type t) const
 {
-	return _atomFactory[t];
+	return t < _atomFactory.size() ? _atomFactory[t] : nullptr;
 }
 
 ClassServer::Validator* ClassServer::getValidator(Type t) const
 {
-	return _validator[t];
+	return t < _validator.size() ? _validator[t] : nullptr;
 }
 
 Handle ClassServer::factory(const Handle& h) const
@@ -251,48 +130,8 @@ Handle ClassServer::factory(const Handle& h) const
 	return h;
 }
 
-Type ClassServer::getNumberOfClasses() const
-{
-    return nTypes;
-}
-
-bool ClassServer::isA_non_recursive(Type type, Type parent) const
-{
-    std::lock_guard<std::mutex> l(type_mutex);
-    if ((type >= nTypes) || (parent >= nTypes)) return false;
-    return inheritanceMap[parent][type];
-}
-
-bool ClassServer::isDefined(const std::string& typeName) const
-{
-    std::lock_guard<std::mutex> l(type_mutex);
-    return name2CodeMap.find(typeName) != name2CodeMap.end();
-}
-
-Type ClassServer::getType(const std::string& typeName) const
-{
-    std::lock_guard<std::mutex> l(type_mutex);
-    std::unordered_map<std::string, Type>::const_iterator it = name2CodeMap.find(typeName);
-    if (it == name2CodeMap.end()) {
-        return NOTYPE;
-    }
-    return it->second;
-}
-
-const std::string& ClassServer::getTypeName(Type type) const
-{
-    static std::string nullString = "*** Unknown Type! ***";
-
-    if (nTypes <= type) return nullString;
-
-    std::lock_guard<std::mutex> l(type_mutex);
-    const std::string* name = _code2NameMap[type];
-    if (name) return *name;
-    return nullString;
-}
-
 ClassServer& opencog::classserver()
 {
-    static std::unique_ptr<ClassServer> instance(new ClassServer());
+    static std::unique_ptr<ClassServer> instance(new ClassServer(nameserver()));
     return *instance;
 }
