+++ opencog/atoms/base/ClassServer.h
@@ -30,8 +30,9 @@
 #include <vector>
 
 #include <opencog/util/sigslot.h>
-#include <opencog/atoms/base/types.h>
-#include <opencog/atoms/base/atom_types.h>
+#include <opencog/atoms/proto/types.h>
+#include <opencog/atoms/proto/atom_types.h>
+#include <opencog/atoms/proto/NameServer.h>
 #include <opencog/atoms/base/Handle.h>
 
 namespace opencog
@@ -40,11 +41,8 @@ namespace opencog
  *  @{
  */
 
-typedef SigSlot<Type> TypeSignal;
-
 /**
- * This class keeps track of the complete protoatom (value and atom)
- * class hierarchy. It also provides factories for those atom types
+ * This class provides factories for those atom types
  * that have non-trivial C++ objects behind them.
  */
 class ClassServer
@@ -60,9 +58,7 @@ public:
 private:
 
     /** Private default constructor for this class to make it a singleton. */
-    ClassServer();
-
-    std::set< std::string > _loaded_modules;
+    ClassServer(const NameServer &);
 
     /* It is very tempting to make the type_mutex into a reader-writer
      * mutex. However, it appears that this is a bad idea: reader-writer
@@ -72,35 +68,18 @@ private:
      * to be a lose-lose proposition. See the Anthony Williams post here:
      * http://permalink.gmane.org/gmane.comp.lib.boost.devel/211180
      */
-    mutable std::mutex type_mutex;
+    mutable std::mutex factory_mutex;
 
-    Type nTypes;
-    Type _maxDepth;
-
-    std::vector< std::vector<bool> > inheritanceMap;
-    std::vector< std::vector<bool> > recursiveMap;
-    std::unordered_map<std::string, Type> name2CodeMap;
-    std::vector<const std::string*> _code2NameMap;
     mutable std::vector<AtomFactory*> _atomFactory;
     mutable std::vector<Validator*> _validator;
-    std::vector<int> _mod;
-    TypeSignal _addTypeSignal;
-
-    void setParentRecursively(Type parent, Type type, Type& maxd);
 
     void spliceFactory(Type, AtomFactory*);
+
+    const NameServer & _nameServer;
 
 public:
     /** Gets the singleton instance (following meyer's design pattern) */
     friend ClassServer& classserver();
-
-    bool beginTypeDecls(const char *);
-    void endTypeDecls(void);
-    /**
-     * Adds a new atom type with the given name and parent type.
-     * Return a numeric value that is assigned to the new type.
-     */
-    Type declType(const Type parent, const std::string& name);
 
     /**
      * Declare a factory for an atom type.
@@ -119,171 +98,6 @@ public:
      * same type.
      */
     Handle factory(const Handle&) const;
-
-    /** Provides ability to get type-added signals.
-     * @warning methods connected to this signal must not call
-     * ClassServer::addType or things will deadlock.
-     */
-    TypeSignal& typeAddedSignal();
-
-    /**
-     * Given the type `type`, get the immediate descendents.
-     * This is NOT recursive, only the first generation is returned.
-     * Stores the children types on the OutputIterator 'result'.
-     * Returns the number of children types.
-     */
-    template<typename OutputIterator>
-    unsigned long getChildren(Type type, OutputIterator result) const
-    {
-        unsigned long n_children = 0;
-        for (Type i = 0; i < nTypes; ++i) {
-            if (inheritanceMap[type][i] and (type != i)) {
-                *(result++) = i;
-                n_children++;
-            }
-        }
-        return n_children;
-    }
-
-    /**
-     * Given the type `type`, get the immediate parents.
-     * This is NOT recursive, only the first generation is returned.
-     * Stores the parent types on the OutputIterator 'result'.
-     * Returns the number of parent types.
-     */
-    template<typename OutputIterator>
-    unsigned long getParents(Type type, OutputIterator result) const
-    {
-        unsigned long n_parents = 0;
-        for (Type i = 0; i < nTypes; ++i) {
-            if (inheritanceMap[i][type] and (type != i)) {
-                *(result++) = i;
-                n_parents++;
-            }
-        }
-        return n_parents;
-    }
-
-    /**
-     * Given the type `type`, get all of the descendents.  This is
-     * recursive, that is, children of the children are returned.
-     * Stores the children types on the OutputIterator 'result'.
-     * Returns the number of children types.
-     */
-    template <typename OutputIterator>
-    unsigned long getChildrenRecursive(Type type, OutputIterator result) const
-    {
-        unsigned long n_children = 0;
-        for (Type i = 0; i < nTypes; ++i) {
-            if (recursiveMap[type][i] and (type != i)) {
-                *(result++) = i;
-                n_children++;
-            }
-        }
-        return n_children;
-    }
-
-    template <typename Function>
-    void foreachRecursive(Function func, Type type) const
-    {
-        for (Type i = 0; i < nTypes; ++i) {
-            if (recursiveMap[type][i]) (func)(i);
-        }
-    }
-
-    /**
-     * Returns the total number of classes in the system.
-     *
-     * @return The total number of classes in the system.
-     */
-    Type getNumberOfClasses() const;
-
-    /**
-     * Returns whether a given class is assignable from another.
-     * This is the single-most commonly called method in this class.
-     *
-     * @param super Super class.
-     * @param sub Subclass.
-     * @return Whether a given class is assignable from another.
-     */
-    bool isA(Type sub, Type super) const
-    {
-        /* Because this method is called extremely often, we want
-         * the best-case fast-path for it.  Since updates are extremely
-         * unlikely after initialization, we use a multi-reader lock,
-         * and don't care at all about writer starvation, since there
-         * will almost never be writers. However, see comments above
-         * about multi-reader-locks -- we are not using them just right
-         * now, because they don't seem to actually help.
-         *
-         * Currently, this lock accounts for 2% or 3% performance
-         * impact on atom insertion into atomspace.  The unit tests
-         * don't need it to pass.  Most users probably dont need it
-         * at all, because most type creation/update happens in
-         * shared-lib ctors, which mistly should be done by the time
-         * that this gets called. How big a price do you want to pay
-         * for avoiding a possible crash on a shared-lib load while
-         * also running some multi-threaded app?
-         */
-        // std::lock_guard<std::mutex> l(type_mutex);
-        if ((sub >= nTypes) || (super >= nTypes)) return false;
-        return recursiveMap[super][sub];
-    }
-
-    bool isA_non_recursive(Type sub, Type super) const;
-
-    /**
-     * Returns true if given class is a Value.
-     *
-     * @param t class.
-     * @return Whether a given class is Value.
-     */
-    bool isValue(Type t) const { return isA(t, VALUE); }
-
-    /**
-     * Returns true if given class is a valid atom type.
-     *
-     * @param t class.
-     * @return Whether a given class is an atom.
-     */
-    bool isAtom(Type t) const { return isA(t, ATOM); }
-
-    /**
-     * Returns true if given class is a Node.
-     *
-     * @param t class.
-     * @return Whether a given class is Node.
-     */
-    bool isNode(Type t) const { return isA(t, NODE); }
-
-    /**
-     * Returns true if given class is a Link.
-     *
-     * @param t class.
-     * @return Whether a given class is Link.
-     */
-    bool isLink(Type t) const { return isA(t, LINK); }
-
-    /**
-     * Returns whether a class with name 'typeName' is defined.
-     */
-    bool isDefined(const std::string& typeName) const;
-
-    /**
-     * Returns the type of a given class.
-     *
-     * @param typeName Class type name.
-     * @return The type of a givenn class.
-     */
-    Type getType(const std::string& typeName) const;
-
-    /**
-     * Returns the string representation of a given atom type.
-     *
-     * @param type Atom type code.
-     * @return The string representation of a givenn class.
-     */
-    const std::string& getTypeName(Type type) const;
 };
 
 ClassServer& classserver();
